+++ lib/i18n/tasks/scanners/erb_ast_scanner.rb
@@ -111,7 +111,8 @@ module I18n::Tasks::Scanners
             path,
             content,
             start + occurrence.pos,
-            raw_key: occurrence.raw_key
+            raw_key: occurrence.raw_key,
+            candidate_keys: occurrence.candidate_keys
           )
         ]
       end
@@ -128,7 +129,8 @@ module I18n::Tasks::Scanners
             path,
             content,
             start + (code.index(key) || occurrence.pos),
-            raw_key: occurrence.raw_key
+            raw_key: occurrence.raw_key,
+            candidate_keys: occurrence.candidate_keys
           )
         ]
       end