+++ lib/i18n/tasks/scanners/prism_scanners/visitor.rb
@@ -95,6 +95,8 @@ module I18n::Tasks::Scanners::PrismScanners
         @current_class&.private_methods!
       when :t, :t!, :translate, :translate!
         args, kwargs = process_arguments(node)
+        # Do not process other receivers than I18n, e.g. Service.translate(:key)
+        return if node.receiver.present? && node.receiver.name != :I18n
         parent.add_translation_call(
           TranslationCall.new(
             node: node,
@@ -268,7 +270,7 @@ module I18n::Tasks::Scanners::PrismScanners
       # We need to check for `node.receiver` since node is the `human` call
       model_name = if current_class.present? && rails_model_method_called_on_current_class?(node.receiver)
         current_class.path.flatten.map!(&:underscore).join(".")
-      elsif node.receiver.present?
+      elsif node.receiver&.receiver&.type == :constant_read_node
         node.receiver&.receiver&.name&.to_s&.underscore
       end
 
@@ -289,6 +291,9 @@ module I18n::Tasks::Scanners::PrismScanners
           node: node,
           receiver: nil,
           key: [:activerecord, :models, model_name, count_key].join("."),
+          candidate_keys: [
+            [:activerecord, :models, model_name].join(".")
+          ],
           parent: parent,
           options: kwargs
         )
@@ -327,6 +332,7 @@ module I18n::Tasks::Scanners::PrismScanners
         TranslationCall.new(
           node: node,
           key: key,
+          candidate_keys: Array([:attributes, array_args.first].join(".")),
           receiver: nil,
           parent: parent,
           options: {}